PCM audio codec for pro res file
As Davinci resolve has limited codec suppor so it does with audio. The addition of the audio codec is to solve render and sound issues using it. Otherwise with unsupported audio codecs in the video footage resolve will never start render the video and will get stucked with a partially loked user interface.